I went down quite the rabbit hole to get this beautiful effect in place. Now, highlighted text has this awesome pop effect. The rest of the text starts transparent and fades in, while the highlighted text's shadow fades away and they blend into a regular style.
It's really quite simple in concept, but it took a while to realize the problem in ReactMarkdown's setup to apply styles to a parent p element which inherently applies to all children, and whose effect cannot be overriden with 'clever' css. the solution was to force the p children all siblings, and apply the styles to the default children and leave the other children (eg: bold, italic, or whatever else) unaffected. aah.